Fixed PointAtTargetSchema a bit
added pointAtTarget to TeleopDrive
It should switch between pointing at a target and driving normally when the Y button is pressed. TeleopRotateSchema and PointAtTargetSchema changed to only be active one at a time (Weight is 0 when not active).
In PointAtTarget:
replaced atan with atan2. Added some wrapping. Still spins when activated for some reason.